Установит prop-types и добавит валидацию пропсов для компонент
