+++ .eslintrc.yml
@@ -16,7 +16,3 @@ plugins: ['react']
 settings:
   react:
     version: '16'
-
-rules: {
-  'react/prop-types': 0
-}
